Preserve decimal precision in comments-based grade display
Match the tables-based fix: use (float) instead of (int) so grades
like 60.98% are not truncated to 60%.

+++ includes/internal/services/class-comments-based-grading-listing-service.php
@@ -76,7 +76,7 @@ class Comments_Based_Grading_Listing_Service implements Grading_Listing_Service_
 				(int) $comment->user_id,
 				(int) $comment->comment_post_ID,
 				$comment->comment_date,
-				'' !== $grade_value ? (int) $grade_value : null
+				'' !== $grade_value ? (float) $grade_value : null
 			);
 		}
